+++ src/pages/productListPage/productListPage.styled.ts
@@ -1,0 +1,6 @@
+import Styled from '@emotion/styled';
+
+export const MiddleContainer = Styled.div`
+  height:calc(100vh - 64px);
+  overflow-y: scroll;
+`;
